fix(kb): extract syncWorkspace to leaf module — keep next/headers out of WS bundle
The edit_c4_diagram tool bundles c4-writer into the WS/custom server
(via cc-dispatcher). c4-writer imported syncWorkspace from
kb-route-helpers, which top-level imports @/lib/supabase/server
(next/headers). esbuild inlines local imports (even dynamic) into the
single dev-server.mjs, so next/headers entered the server bundle and Node
ESM failed to resolve the bare specifier → custom server crashed at
startup → e2e WebServer failed.

Extract syncWorkspace + classifyGitSyncError + selfHealNonFastForward +
SyncWorkspaceResult into server/workspace-sync.ts (deps: observability +
lazy git-auth + types only — no supabase/server). kb-route-helpers
re-exports syncWorkspace/SyncWorkspaceResult so its callers are
unaffected; c4-writer imports from the leaf. Verified: dev-server esbuild
bundle has 0 next/headers; full suite 8548 pass.
